Favicon tweak
Subbed out the <link> for the favicon to point to a favicon.ico file at the root since some old versions of Firefox (which this should still work on) look for it there.

Left the original Icon.png file in place in case it needs to be used elsewhere, though I don't think it currently is.